Make function q_insert_tail works properly
In order to make the insertion at tail operate in O(1) time,
a new field named tail which is in type of list_ele_t** is
introduced to queue_t. The initial value (while queue is empty)
of tail would be the address of queue_t.head.
The reason why define queue_t.tail as type of list_ele_t** is to
make code more concise, which also makes some checks on edge
condition unnecessary.