Make function q_sort works properly
The algorithm being choosed here is merge sort, in order to reach
O(nlog) time complexity. To fulfill it, two static functions,
divide and merge, are introduced into the program.
Function divide will cut the given queue a into two subqueues, and
the other subqueue will be place into queue b. Since this
function only be called in function q_sort, which will first check
if given queue has at lest 2 elements, function divide keeps this
as its assumption, and will not check for edge conditions such as
NULL queue or empty queue.
Function merge will assume that the given queues are both in
acensing order, and it will merge b into a in acensing order, too.
And with the same assumption as above, this function will not check
the edge conditions.